+++ core/src/options/from_variant.rs
@@ -9,6 +9,9 @@ use {FromMeta, Result};
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct FromVariantOptions {
     pub base: OuterFrom,
+    /// The field on the deriving struct into which the discriminant expression
+    /// should be placed by the derived `FromVariant` impl.
+    pub discriminant: Option<Ident>,
     pub fields: Option<Ident>,
     pub supports: Option<DataShape>,
 }
@@ -17,6 +20,7 @@ impl FromVariantOptions {
     pub fn new(di: &DeriveInput) -> Result<Self> {
         (FromVariantOptions {
             base: OuterFrom::start(di),
+            discriminant: Default::default(),
             fields: Default::default(),
             supports: Default::default(),
         })
@@ -30,6 +34,7 @@ impl<'a> From<&'a FromVariantOptions> for FromVariantImpl<'a> {
         FromVariantImpl {
             base: (&v.base.container).into(),
             ident: v.base.ident.as_ref(),
+            discriminant: v.discriminant.as_ref(),
             fields: v.fields.as_ref(),
             attrs: v.base.attrs.as_ref(),
             attr_names: &v.base.attr_names,
@@ -60,6 +65,10 @@ impl ParseData for FromVariantOptions {
             .as_ref()
             .map(|v| v.as_str())
         {
+            Some("discriminant") => {
+                self.discriminant = field.ident.clone();
+                Ok(())
+            }
             Some("fields") => {
                 self.fields = field.ident.clone();
                 Ok(())